+++ types/gen/go/api/v1/api.pb.validate.go
@@ -4698,6 +4698,48 @@ func (m *Step) validate(all bool) error {
 			}
 		}
 
+	case *Step_Confluence:
+		if v == nil {
+			err := StepValidationError{
+				field:  "Config",
+				reason: "oneof value cannot be a typed-nil",
+			}
+			if !all {
+				return err
+			}
+			errors = append(errors, err)
+		}
+		oneofConfigPresent = true
+
+		if all {
+			switch v := interface{}(m.GetConfluence()).(type) {
+			case interface{ ValidateAll() error }:
+				if err := v.ValidateAll(); err != nil {
+					errors = append(errors, StepValidationError{
+						field:  "Confluence",
+						reason: "embedded message failed validation",
+						cause:  err,
+					})
+				}
+			case interface{ Validate() error }:
+				if err := v.Validate(); err != nil {
+					errors = append(errors, StepValidationError{
+						field:  "Confluence",
+						reason: "embedded message failed validation",
+						cause:  err,
+					})
+				}
+			}
+		} else if v, ok := interface{}(m.GetConfluence()).(interface{ Validate() error }); ok {
+			if err := v.Validate(); err != nil {
+				return StepValidationError{
+					field:  "Confluence",
+					reason: "embedded message failed validation",
+					cause:  err,
+				}
+			}
+		}
+
 	default:
 		_ = v // ensures v is used
 	}